+++ src/tightdb/column_linklist.hpp
@@ -53,9 +53,9 @@ public:
 
     LinkViewRef get_link_view(std::size_t row_ndx);
 
+    void erase(std::size_t, bool) TIGHTDB_OVERRIDE;
+    void move_last_over(std::size_t, std::size_t) TIGHTDB_OVERRIDE;
     void clear() TIGHTDB_OVERRIDE;
-    void move_last_over(std::size_t ndx) TIGHTDB_OVERRIDE;
-    void erase(std::size_t ndx, bool is_last) TIGHTDB_OVERRIDE;
 
     // ColumnLinkBase overrides
     void set_target_table(TableRef table) TIGHTDB_OVERRIDE;