+++ src/tightdb/row.cpp
@@ -7,15 +7,18 @@ using namespace tightdb;
 
 void RowBase::attach(Table* table, size_t row_ndx)
 {
-    table->register_row_accessor(this); // Throws
-    m_table.reset(table);
-    m_row_ndx = row_ndx;
+    if (table) {
+        table->register_row_accessor(this); // Throws
+        m_table.reset(table);
+        m_row_ndx = row_ndx;
+    }
 }
 
 void RowBase::reattach(Table* table, size_t row_ndx)
 {
     if (m_table.get() != table) {
-        table->register_row_accessor(this); // Throws
+        if (table)
+            table->register_row_accessor(this); // Throws
         if (m_table)
             m_table->unregister_row_accessor(this);
         m_table.reset(table);