+++ release_notes.md
@@ -39,6 +39,8 @@ C++ (core)
 * Fixed bugs in Table::insert_column() and Table::remove_column() causing crash or undefined behaviour.
 * Fixed corruption bug when a string enumeration column follows a column with attached search index (index flavor mixup).
 * Fixed in Array::erase() causing crash in certain row insertion scenarios.
+* Fixed bug in enumerated strings column (corruption was possible when inserting default values).
+* Fixed bug in Table::update_from_parent() causing a crash if Group::commit() in presence of generated subtable accessor.
 
 ### API breaking changes:
 
@@ -49,6 +51,8 @@ C++ (core)
 * Added support for links and lists of links as column types, to enable relationships between tables.
 * Adding Table::get_index_in_parent() and Group::get_table(std::size_t table_ndx). They were needed for implicit transactions.
 * Table::get_parent_table() can now also return the index of the column in the parent.
+* Support for row accessors.
+* Table, row, and descriptor accessors are now generally retained and properly adjusted when the parent table is modified.
 
 -----------